% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular.R
\name{calcular}
\alias{calcular}
\title{Obter abundancia relativa e categorias}
\usage{
calcular(a)
}
\arguments{
\item{a}{data.frame contendo na coluna 1 o nome dos táxons e na coluna 2 a abundância absoluta.}
}
\description{
Esta função obtém a abundancia relativa os categoriza.
}
