% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{color_table}
\alias{color_table}
\title{Add color css to table ?}
\usage{
color_table(table_html, columns, colors)
}
\arguments{
\item{table_html}{(\code{character} of length 1)}

\item{columns}{(\code{integer})}

\item{colors}{?}
}
\description{
Add color css to table ?
}

