% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{depth}
\alias{depth}
\title{Get max depth of a recursive list}
\usage{
depth(this)
}
\arguments{
\item{this}{(\code{list})}
}
\description{
\url{http://stackoverflow.com/questions/13432863/determine-level-of-nesting-in-r}
}
