% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.and.louvain.R
\name{hclust.and.louvain}
\alias{hclust.and.louvain}
\title{Do both hierarchical clustering with hclust
And also louvain clustering with igraph}
\usage{
hclust.and.louvain(
  data,
  nearest.neighbors = 25,
  dist.method = "euclidean",
  hclust.method = "average",
  knn.graph = NULL,
  row.name = "Gene"
)
}
\arguments{
\item{data}{A matrix or dataframe to be clustered on rows}

\item{nearest.neighbors}{The number of neighbors to keep for the knn.graph}

\item{dist.method}{The distance method (see 'dist')}

\item{hclust.method}{The hclust method (see 'hclust)}

\item{knn.graph}{An igraph graph object which will be used instead of k-NN inverse distance for louvain clustering}

\item{row.name}{The name used for rows in the louvain dataframe}
}
\value{
An hclust object with an additional 'louvain' attribute (that )
}
\description{
Do both hierarchical clustering with hclust
And also louvain clustering with igraph
}
