% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, Gresham Lab!}
\usage{
hello()
}
\description{
This is an example function named 'hello'
which prints 'Hello, Gresham Lab!'.
}
\details{
You can contribute to this package at:

  https://github.com/GreshamLab/labtools
}
