% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_from_snapgene_features.R
\name{make_gff_from_snap}
\alias{make_gff_from_snap}
\title{Make a gff from a snap gene file}
\usage{
make_gff_from_snap(file_path, chromosome, feature_source, output)
}
\arguments{
\item{file_path}{the file_path is the name of the csv which has the snapgene feature output}

\item{chromosome}{is the chromosome the features are on.}

\item{feature_source}{is what you want in the source column}

\item{output}{is what you want your output gff to be called (should be .gff)}
}
\value{
A file formatted as a gff
}
\description{
this is a function that takes a csv with features generated from snapgene, and converts it into a gff format
when exporting features from snapgene, export all feature data (csv should have 5 columns)
}
\author{
Grace Avecilla \email{ga824@nyu.edu }
}
