% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcr.R
\name{pcr_table}
\alias{pcr_table}
\title{Creates table for PCR recipie}
\usage{
pcr_table(count, additives = c(DNA = 1), additive_concentration = rep("",
  length(additives)))
}
\arguments{
\item{count}{(\code{integer} of length 1) The number of PCR reactions.}

\item{additives}{(named \code{numeric}) The volume non-mastermix ingredients that will be added.}

\item{additive_concentration}{(named \code{numeric}) The concentration of non-mastermix ingredients.}
}
\description{
Creates a markdown table for a PCR master mix recipe.
}
\examples{
pcr_table(count=8, additives=c(DNA=1, other=7))
}
\keyword{PCR}

