% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_msdial_data.R
\name{read_msdial}
\alias{read_msdial}
\title{Read in MS-DIAL exported *.txt file and clean it}
\usage{
read_msdial(file, type = "gcms")
}
\arguments{
\item{file}{The *.txt file exported from MS-DIAL.}

\item{type}{Type of the data analyed by MS-DIAL, can be either gcms or lcms.
lcms is not currently implemented.}
}
\value{
A cleaned table with most important information reserved.
}
\description{
Read in MS-DIAL exported *.txt file and clean it
}
