% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_nanodrop_spectrum_tsv}
\alias{read_nanodrop_spectrum_tsv}
\title{Parses the spectrum output of the Nanodrop spectrophotometer}
\usage{
read_nanodrop_spectrum_tsv(path, average = TRUE, melt = FALSE)
}
\arguments{
\item{path}{(\code{character} of length 1) The file path to the .tsv output file of
Nanodrop measurments.}

\item{average}{(\code{logical} of length 1) If \code{TRUE}, all numeric data for
measurments with the same ID are averaged.}

\item{melt}{(\code{logical} of length 1) If \code{TRUE}, melt data for use in ggplot2.}
}
\description{
Parses the spectrum output of the Nanodrop spectrophotometer into a dataframe.
}
\keyword{Nanodrop}
\keyword{nanodrop}
\keyword{spectrophotometer}
\keyword{spectrum}
