% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_nanodrop_tsv}
\alias{read_nanodrop_tsv}
\title{Parses the output of the Nanodrop spectrophotometer}
\usage{
read_nanodrop_tsv(path, average = TRUE)
}
\arguments{
\item{path}{(\code{character} of length 1) The file path to the .tsv output file of
Nanodrop measurments.}

\item{average}{(\code{logical} of length 1) If \code{TRUE}, all numeric data for
measurments with the same ID are averaged.}
}
\description{
Parses the output of the Nanodrop spectrophotometer into a dataframe.
}
\keyword{Nanodrop}
\keyword{nanodrop}
\keyword{spectrophotometer}

