% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_err.R
\name{std_err}
\alias{std_err}
\title{Compute the standard error for a vector of numbers.}
\usage{
std_err(x)
}
\arguments{
\item{x}{a vector of values.}
}
\value{
The standard error of \code{x}.
}
\description{
Compute the standard error for a vector of numbers.
}
\examples{
std_err(c(4,4,5,5,6,6))
std_err(rnorm(100))
}
\seealso{
\url{https://github.com/davidgresham/davidtools}
}
