% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_common_to_systematic.R
\name{yeast common to systematic}
\alias{yeast common to systematic}
\title{yeast_common_to_systematic}
\arguments{
\item{name_vec}{a vector of common names.}
}
\value{
A vector of systematic names. If no systematic name is found, returns common name
}
\description{
yeast_common_to_systematic
}
\examples{
yeast_common_to_systematic("GAP1")

}
