% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_r64_to_systematic.R
\name{yeast R64 ID systematic}
\alias{yeast R64 ID systematic}
\title{yeast_r64_to_systematic}
\arguments{
\item{name_vec}{a vector of GCF_000146045.2_R64_genomic IDs (format: ID=cdsNN).}
}
\value{
A vector of systematic names. If no systematic name is found, returns GCF_000146045.2_R64_genomic ID
}
\description{
yeast_r64_to_systematic
}
\examples{
yeast_r64_to_systematic("ID=cds2818")

}
