% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_sgd_to_common.R
\name{yeast SGD ID to common}
\alias{yeast SGD ID to common}
\title{yeast_sgd_to_common}
\arguments{
\item{name_vec}{a vector of SGD IDs (format: SGD:SNNNNNNNNN).}
}
\value{
A vector of common names. If no common name is found, returns SGD ID
}
\description{
yeast_sgd_to_common
}
\examples{
yeast_sgd_to_common("SGD:S000005086")

}
