% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_systematic_to_sgd.R
\name{yeast systematic to SGD ID}
\alias{yeast systematic to SGD ID}
\title{yeast_systematic_to_sgd}
\arguments{
\item{name_vec}{a vector of systematic names.}
}
\value{
A vector of SGD IDs (format: SGD:SNNNNNNNNN). If no SGD ID is found, returns systematic name
}
\description{
yeast_systematic_to_sgd
}
\examples{
yeast_systematic_to_sgd("YHL016C")

}
