% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_common_to_sgd.R
\name{yeast_common_to_sgd}
\alias{yeast_common_to_sgd}
\title{yeast_common_to_sgd}
\arguments{
\item{name_vec}{a vector of common names.}
}
\value{
A vector of SGD IDs (format: SGD:SNNNNNNNNN) If no SGD ID is found, returns common name
}
\description{
yeast_common_to_sgd
}
\examples{
yeast_common_to_sgd("GAL4")

}
