% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_gene_names.R
\docType{data}
\name{yeast_gene_names}
\alias{yeast_gene_names}
\title{Gene names for 5985 yeast genes.}
\format{
A data frame with 5985 rows and 4 variables:
\describe{
  \item{Systematic_name}{systematic name}
  \item{Common_name}{common name}
  \item{GCF_000146045.2_R64_genomic_ID}{Genbank name}
  \item{SGD_ID}{SGD ID}
  ...
}
}
\source{
\url{https://www.yeastgenome.org/}
}
\usage{
yeast_gene_names
}
\description{
A dataset containing the names of 5,895
yeast genes.
}
\keyword{datasets}
