% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_sgd_to_systematic.R
\name{yeast_sgd_to_systematic}
\alias{yeast_sgd_to_systematic}
\title{yeast_sgd_to_systematic}
\arguments{
\item{name_vec}{a vector of SGD IDs (format: SGD:SNNNNNNNNN).}
}
\value{
A vector of systematic names. If no systematic name is found, returns SGD ID
}
\description{
yeast_sgd_to_systematic
}
\examples{
yeast_sgd_to_systematic("SGD:S000000979")

}
