% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_database_for_msfinder.R
\name{export4msfinder}
\alias{export4msfinder}
\title{Export structure database for MS-FINDER}
\usage{
export4msfinder(
  input,
  cas_col = FALSE,
  output = "structure data for msfinder.txt"
)
}
\arguments{
\item{input}{The path of the list of chemicals. The file can be either in txt,
csv, xlsl, or xls form, but the chemical name should be in the first column. #'
for example, "c:/data/list_of_chemicals.txt"}

\item{cas_col}{If you have CAS number for each chemical, then specify in which
column the CAS number locates. It will be used for extracting chemical information
from Pubchem. This is recommended. If no CAS number is available, just levea it,
and it will use the chemical name for extraction.}

\item{output}{The path where the structure database will be stored. It must be
in txt format. For example, "c:/data/structure_database_for_msfinder.txt"}
}
\description{
\code{export4msfinder} provides a way to convert any list of chemicals into a
structure database that can be used by MS-FINDER.
}
