% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_systematic_to_common.R
\name{yeast systematic to common}
\alias{yeast systematic to common}
\title{yeast_systematic_to_common}
\arguments{
\item{name_vec}{a vector of systematic names.}
}
\value{
A vector of common names. If no common name is found, returns systematic name
}
\description{
yeast_systematic_to_common
}
\examples{
yeast_systematic_to_common("YFL039C")

}
